+++ esphome/components/esp_audio_stack/speaker/esp_audio_stack_speaker.cpp
@@ -27,14 +27,13 @@ void ESPAudioStackSpeaker::setup() {
 
   // Forward frame-played notifications from I2S audio task to mixer callbacks.
   // Without this, mixer source speakers can't track pending_playback_frames.
-  this->parent_->add_speaker_output_callback([this](uint32_t frames, int64_t timestamp) {
-    this->audio_output_callback_.call(frames, timestamp);
-  });
+  this->parent_->add_speaker_output_callback(
+      [this](uint32_t frames, int64_t timestamp) { this->audio_output_callback_.call(frames, timestamp); });
 }
 
 void ESPAudioStackSpeaker::dump_config() {
   ESP_LOGCONFIG(TAG, "ESP Audio Stack Speaker:");
-  ESP_LOGCONFIG(TAG, "  Sample Rate: %u Hz", this->parent_->get_sample_rate());
+  ESP_LOGCONFIG(TAG, "  Sample Rate: %u Hz", (unsigned) this->parent_->get_sample_rate());
   ESP_LOGCONFIG(TAG, "  Bits Per Sample: 16");
   ESP_LOGCONFIG(TAG, "  Channels: %u", (unsigned) this->parent_->get_speaker_channels());
   ESP_LOGCONFIG(TAG, "  Buffer Size: %u bytes", (unsigned) this->parent_->get_speaker_buffer_size());
@@ -58,8 +57,8 @@ void ESPAudioStackSpeaker::start() {
 
   // Idempotent: register listener only once per stream session.
   bool expected = false;
-  if (!this->listener_registered_.compare_exchange_strong(
-          expected, true, std::memory_order_acq_rel, std::memory_order_relaxed))
+  if (!this->listener_registered_.compare_exchange_strong(expected, true, std::memory_order_acq_rel,
+                                                          std::memory_order_relaxed))
     return;
 
   if (xSemaphoreTake(this->active_listeners_semaphore_, 0) != pdTRUE) {
@@ -94,12 +93,9 @@ void ESPAudioStackSpeaker::finish() {
   this->enable_loop_soon_any_context();
 }
 
-size_t ESPAudioStackSpeaker::play(const uint8_t *data, size_t length) {
-  return this->play(data, length, 0);
-}
+size_t ESPAudioStackSpeaker::play(const uint8_t *data, size_t length) { return this->play(data, length, 0); }
 
-size_t ESPAudioStackSpeaker::play(const uint8_t *data, size_t length,
-                                    TickType_t ticks_to_wait) {
+size_t ESPAudioStackSpeaker::play(const uint8_t *data, size_t length, TickType_t ticks_to_wait) {
   if (this->is_failed()) {
     return 0;
   }
@@ -123,9 +119,7 @@ size_t ESPAudioStackSpeaker::play(const uint8_t *data, size_t length,
   return written;
 }
 
-bool ESPAudioStackSpeaker::has_buffered_data() const {
-  return this->parent_->get_speaker_buffer_available() > 0;
-}
+bool ESPAudioStackSpeaker::has_buffered_data() const { return this->parent_->get_speaker_buffer_available() > 0; }
 
 void ESPAudioStackSpeaker::set_volume(float volume) {
   if (!std::isfinite(volume) || volume < 0.0f) {
